requirements.txt
gradio
google-cloud-aiplatform[evaluation]
plotly
pandas